REFACTOR: 이메일 인증 구현
코드 리팩토링
JwtToken 인증방식 명확성 향상
불필요한 코드 삭제
