+++ src/main/java/backend/dev/user/repository/UserRepository.java
@@ -1,19 +1,15 @@
 package backend.dev.user.repository;
 
 import backend.dev.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface UserRepository extends JpaRepository<User,String> {
-
+    @Query("select u from User u where u.email = :email")
     Optional<User> findByEmail(String email);
 
     @Query("select u from User u join fetch u.oauthList o where o.provider = :provider and o.providerId = :providerId")
     Optional<User> findByProviderAndProviderId(String provider, String providerId);
-
 
 }
